Move ValidationError definition into Model
This allows the PersistenceModel to share these definitions for the
purpose of serializing HmdaFileValidator messages.
