+++ api/src/main/scala/hmda/api/http/ValidationErrorConverter.scala
@@ -3,6 +3,7 @@ package hmda.api.http
 import hmda.api.model._
 import hmda.model.edits.EditMetaDataLookup
 import hmda.model.fi.{ HmdaFileRow, HmdaRowError }
+import hmda.model.validation.ValidationError
 import hmda.persistence.processing.HmdaFileValidator.HmdaFileValidationState
 import hmda.validation.engine._
 import spray.json.{ JsNumber, JsObject, JsString, JsValue }
