+++ api/src/main/scala/hmda/api/http/institutions/submissions/SubmissionEditPaths.scala
@@ -13,6 +13,7 @@ import hmda.api.http.{ HmdaCustomDirectives, ValidationErrorConverter }
 import hmda.api.model._
 import hmda.api.protocol.processing.{ ApiErrorProtocol, EditResultsProtocol, InstitutionProtocol }
 import hmda.model.fi.{ Submission, SubmissionId, SubmissionStatus }
+import hmda.model.validation.{ Macro, Quality, ValidationErrorType }
 import hmda.persistence.messages.CommonMessages.GetState
 import hmda.persistence.HmdaSupervisor.{ FindProcessingActor, FindSubmissions }
 import hmda.persistence.institutions.SubmissionPersistence
