+++ api/src/main/scala/hmda/api/protocol/validation/ValidationResultProtocol.scala
@@ -1,6 +1,7 @@
 package hmda.api.protocol.validation
 
 import hmda.api.model.SingleValidationErrorResult
+import hmda.model.validation._
 import hmda.validation.engine._
 import spray.json.{ DefaultJsonProtocol, DeserializationException, JsString, JsValue, RootJsonFormat }
