+++ persistence/src/main/scala/hmda/persistence/processing/HmdaFileValidator.scala
@@ -10,6 +10,7 @@ import hmda.model.fi.SubmissionId
 import hmda.model.fi.lar.LoanApplicationRegister
 import hmda.model.fi.ts.TransmittalSheet
 import hmda.model.institution.Institution
+import hmda.model.validation._
 import hmda.persistence.HmdaSupervisor.{ FindHmdaFiling, FindProcessingActor }
 import hmda.persistence.institutions.InstitutionPersistence
 import hmda.persistence.institutions.InstitutionPersistence.GetInstitution
