+++ validation/src/main/scala/hmda/validation/engine/lar/syntactical/LarSyntacticalEngine.scala
@@ -1,9 +1,9 @@
 package hmda.validation.engine.lar.syntactical
 
 import hmda.model.fi.lar.LoanApplicationRegister
+import hmda.model.validation.Syntactical
 import hmda.validation.api.ValidationApi
 import hmda.validation.context.ValidationContext
-import hmda.validation.engine.Syntactical
 import hmda.validation.engine.lar.LarCommonEngine
 import hmda.validation.rules.ts.syntactical.{ S020, S025 }
 import hmda.validation.rules.lar.syntactical._
