+++ validation/src/main/scala/hmda/validation/engine/lar/validity/LarValidityEngine.scala
@@ -1,9 +1,9 @@
 package hmda.validation.engine.lar.validity
 
 import hmda.model.fi.lar.LoanApplicationRegister
+import hmda.model.validation.Validity
 import hmda.validation.api.ValidationApi
 import hmda.validation.context.ValidationContext
-import hmda.validation.engine.Validity
 import hmda.validation.engine.lar.LarCommonEngine
 import hmda.validation.rules.lar.validity._
